requirements.txt
Babel==2.7.0
Django==2.2.3
django-phonenumber-field==3.0.1
phonenumbers==8.10.14
pytz==2019.1
sqlparse==0.3.0
dj-database-url==0.5.0